kpopit-backend/requirements.txt
blinker==1.9.0
boto3==1.42.28
botocore==1.42.28
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==25.0
pillow==12.0.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
pycountry==24.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
bcrypt==4.2.1
PyJWT==2.10.1
Flask-Limiter==3.9.0
email-validator==2.3.0